把教程部分扔到组件里
